Feature/order

See merge request leson131002/locker!6

+++ LockerService.Application/Common/Services/ICurrentPrincipalService.cs
@@ -6,8 +6,8 @@ public interface ICurrentPrincipalService
 {
     public string? CurrentPrincipal { get; }
     
-    public int? CurrentSubjectId { get; }   
+    public long? CurrentSubjectId { get; }   
     
     public ClaimsPrincipal GetCurrentPrincipalFromToken(string token);
-
+    
 }
